Fix tests by not checking default Content-Type
For some reason the default Content-Type for objects uploaded during CI
testing is `binary/octet-stream`, instead of the standard
`application/octet-stream`.

It doesn't make sense for the upload test to be verifying the default
Content-Type returned when no Content-Type is specified, as that is an
implementation detail of the object store.